pyproject.toml
[project]
name = "opensangamo"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "python-dotenv",
    "httpx",
]
